添加interceptor，统一处理url中的info参数

+++ src/main/java/com/naixwf/article/web/CategoryController.java
@@ -1,13 +1,17 @@
 package com.naixwf.article.web;
 
+import com.naixwf.article.domain.Article;
 import com.naixwf.article.domain.Category;
+import com.naixwf.article.service.ArticleService;
 import com.naixwf.article.service.CategoryService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.stereotype.Controller;
+import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 import javax.annotation.Resource;
 import java.util.List;
@@ -27,9 +31,12 @@ import java.util.Map;
 @Secured({ "ROLE_ADMIN" })
 @Controller
 @RequestMapping("/category")
-public class CategoryController {
+public class CategoryController extends BaseController {
+	private static final Logger LOGGER = LoggerFactory.getLogger(CategoryController.class);
 	@Resource
 	private CategoryService categoryService;
+	@Resource
+	private ArticleService articleService;
 
 	/**
 	 * 分类列表
@@ -48,9 +55,8 @@ public class CategoryController {
 	 */
 	@RequestMapping(value = "/add", method = RequestMethod.POST)
 	public String postAdd(Category category) {
-		//TODO 过滤恶意脚本  防CSRF
 		categoryService.add(category);
-		return "redirect:/category";
+		return "redirect:/category?info" + urlEncode("添加文档分类【%s】成功", category.getCategroyName());
 	}
 
 	/**
@@ -59,7 +65,7 @@ public class CategoryController {
 	@RequestMapping(value = "/edit", method = RequestMethod.POST)
 	public String postEdit(Category category) {
 		categoryService.modify(category);
-		return "redirect:/category";
+		return "redirect:/category?info=" + urlEncode("修改文档分类【%s】成功", category.getCategroyName());
 	}
 
 	/**
@@ -67,9 +73,13 @@ public class CategoryController {
 	 */
 	@RequestMapping(value = "/delete", method = RequestMethod.POST)
 	public String postDelete(Integer categoryId) {
-		//TODO 先查一下有没有article在用这个类别，如果有，则不允许删除
+		List<Article> articleList = articleService.getListByCategoryId(categoryId);
+		if (!CollectionUtils.isEmpty(articleList)) {
+			return "redirect:/category?info=" + urlEncode("该分类下有%d篇文档，不允许删除", articleList.size());
+		}
+
 		categoryService.delete(categoryId);
-		return "redirect:/category";
+		return "redirect:/category?info=" + urlEncode("删除文档分类成功");
 	}
 
 }
